Fatal error fix for class WC_Payment_Gateway
The plugin throws a fatal error in some scenarios as below. As a result
the admin has to manually delete the plugin to activate his site.
http://screencast.com/t/8kGHzxqRNa9E

This happens when:
1. Admin tries to install and activate the Booking plugin while
WooCommerce is not installed or deactivated.
2. Admin tries to update WooCommerce while the booking plugin is active.

It should be noted that the error is intermittent and has been reported
so far by 3 clients.

The fix will do the below:

1. when the user tries to activate booking with WC deactivated, the
plugin is not activated and an error notice is displayed stating
'WooCommerce Booking and Appointment Plugin requires WooCommerce
installed and activate.'

2. If both WC and booking are active, and the user deactivates WC, then
along with WC, booking is also deactivated and the same message is
displayed

3. No fatal error will be thrown from the plugin even if WooCommerce is
deactivated and then reactivated while updating it.

As a part of this fix, the include for all the booking plugin files has
been changed. As a result, it is necessary to conduct a 'DO NOT BREAK'
test for all the features.