Fix for fatal error for BKAP_Payment_Gateway class
Fatal error: Class 'BKAP_Payment_Gateway' not found in
/web/htdocs/www.sightseeinginitaly.com/home/wp-content/plugins/woocommerce/includes/class-wc-payment-gateways.php
on line 104

This fatal error has been reported by a client. However, we're unable to
replicate it on any of the staging sites.

This might be happening as the file is not included when an object of
the class is being created. The fix applied is untested.